+++ src/biz/bokhorst/xprivacy/Meta.java
@@ -62,6 +62,10 @@ public class Meta {
 		mListHook.add(new Hook("calling", "android.intent.action.RESPOND_VIA_MESSAGE", "SEND_RESPOND_VIA_MESSAGE", 18, null, null).doNotify());
 		mListHook.add(new Hook("calling", "android.intent.action.CALL", "CALL_PHONE", 10, null, null).doNotify());
 
+		mListHook.add(new Hook("calling", "SIP.isApiSupported", "USE_SIP", 9, null, null).doNotify());
+		mListHook.add(new Hook("calling", "SIP.isVoipSupported", "USE_SIP", 9, null, null).doNotify());
+		mListHook.add(new Hook("calling", "SIP.newInstance", "USE_SIP", 9, null, null).doNotify());
+
 		mListHook.add(new Hook("clipboard", "addPrimaryClipChangedListener", "", 11, null, null));
 		mListHook.add(new Hook("clipboard", "getPrimaryClip", "", 11, null, null).doNotify());
 		mListHook.add(new Hook("clipboard", "getPrimaryClipDescription", "", 11, null, null).doNotify());